rust-lang: download tarball from GitHub instead of cloning it
- Download the tarball from GitHub as their releases are tagged
It is not necessary to clone sources.

- Makefile changes and reorder things to keep it up with most Makefiles
  in OpenWrt packages repository, e.g., explicitly state that rust-lang
is the host package, it is possible to build it in parallel, etc.

- Fix SPDX License Identifiers

Also, it changes the approach from CONFIG_TARGET_PROFILE, because
this was done only for one device, there is a possibility to use
CONFIG_ARCH, which is taken from target.mk, thus anyone with ARMv7
device can use this in OpenWrt build system. It was working for aarch64
before, thus it is not necessary to explicitly mention it here, but
rather be safe.