angelcam-connector: download tarball from GitHub
This approach is going to be simpler and faster, because it does not
require to have OpenWrt build system. If you want to update your package
because you can download the tarball, check the checksum and put it to the
correct variable. If the hash does not match, then build system does not
proceed anymore to compile this package.

In the previous solution, you need to have a ready OpenWrt build system,
check the PKG_MIRROR_HASH, and even if it is incorrect, then it is still
going to proceed and checkout the sources [1].

[1] https://forum.openwrt.org/t/what-pkg-mirror-hash-or-pkg-hash-means/1602/2