Fix the Rust install dir and simplify the angelcam-connector Makefile (#5)
* Fix the Rust install dir and simplify the angelcam-connector Makefile

Rust will be installed within the build tree (the default location used to
be user's home). As a side-result of this change, some variables common
for rust-lang and angelcam-connector have been moved to
rust-lang/files/rust-env.mk. This file contains also common initialization
for Rust builds that can further simplify Makefiles for Rust apps.

* CR fixes

* Makefile(s) polishing (#6)

* angelcam-connector: Makefile polishing

- Makefile reordering to be more in sync with OpenWrt packages feed
- Sync indentation somewhere were used spaces, somewhere not
- Reordered install section to be more readable

* rust-lang: download tarball from GitHub instead of cloning it

- Download the tarball from GitHub as their releases are tagged
It is not necessary to clone sources.

- Makefile changes and reorder things to keep it up with most Makefiles
  in OpenWrt packages repository, e.g., explicitly state that rust-lang
is the host package, it is possible to build it in parallel, etc.

- Fix SPDX License Identifiers

Also, it changes the approach from CONFIG_TARGET_PROFILE, because
this was done only for one device, there is a possibility to use
CONFIG_ARCH, which is taken from target.mk, thus anyone with ARMv7
device can use this in OpenWrt build system. It was working for aarch64
before, thus it is not necessary to explicitly mention it here, but
rather be safe.

* reforis-angelcam-connector-plugin: remove additional spaces

These spaces were caused by text editor, but many OpenWrt Makefiles does
not have it, so remove them.

* angelcam-connector: download tarball from GitHub

This approach is going to be simpler and faster, because it does not
require to have OpenWrt build system. If you want to update your package
because you can download the tarball, check the checksum and put it to the
correct variable. If the hash does not match, then build system does not
proceed anymore to compile this package.

In the previous solution, you need to have a ready OpenWrt build system,
check the PKG_MIRROR_HASH, and even if it is incorrect, then it is still
going to proceed and checkout the sources [1].

[1] https://forum.openwrt.org/t/what-pkg-mirror-hash-or-pkg-hash-means/1602/2

Co-authored-by: Josef Schlehofer <pepe.schlehofer@gmail.com>